+++ skeleton-app/src/lib/components/organisms/Hands.svelte
